data_generator.py: Changed some names to make easier for beginners to read.
From "The Zen of Python"

    Explicit is better than implicit.

This is simple almost mundane stuff.
Abbreviated names can be figured out,
but they take time to figure out
and can flummox beginners,
so some names were unabbreviated.

One function name, get_email, was arguably wrong.
It did not get (or even make) email.
It made an email _address_, so it was changed to reflect that.